changed code after review
